% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{counts}
\alias{ic}
\alias{exp.anno}
\title{Test data for this package}
\format{
A data frame.

A data frame.

A data frame.
}
\description{
Test data for this package

Test data for this package

Test data for this package
}
\keyword{datasets}
